# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Form Management API
  version: 1.0.0
  description: |
    Comprehensive RESTful API for managing dynamic forms, submissions, and user authentication.

    ## Features
    - Dynamic form creation and management
    - Form submission handling with validation
    - User authentication and authorization
    - Role-based access control (RBAC)
    - Real-time form analytics
    - File upload support
    - Multi-language support

    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

  contact:
    name: API Support
    email: support@formapi.example.com
    url: https://formapi.example.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.formapi.example.com/v1
    description: Production server
  - url: https://staging-api.formapi.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates a user and returns a JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecureP@ssw0rd
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                message: Login successful
                data:
                  user:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    email: user@example.com
                    firstName: John
                    lastName: Doe
                    role: user
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expiresIn: 86400
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
            expiresIn:
              type: integer
              description: Token expiration time in seconds
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            field:
              type: string
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Invalid request data
            error:
              code: VALIDATION_ERROR
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication required
            error:
              code: UNAUTHORIZED
...